+++ src/components/header/Header.type.ts
@@ -1,9 +1,3 @@
-export interface ICategoryProps {
-  img: string
-  description: string
-  type: number
-}
-
 export interface IHeaderProps {
   kebabClick: (isKebabClicked: boolean) => void
 }
