+++ src/components/naviagte-button/NavigateButton.style.tsx
@@ -1,0 +1,25 @@
+import styled from 'styled-components'
+
+export const Navigate = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 4.6rem;
+  gap: 0.4rem;
+  cursor: pointer;
+  span {
+    color: #333;
+    font-family: Pretendard;
+    font-size: 1.2rem;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 1.8rem;
+    white-space: nowrap;
+  }
+`
+export const Icon = styled.img`
+  width: 2.4rem;
+  height: 2.4rem;
+  cursor: pointer;
+`
