+++ src/components/naviagte-button/NavigateButton.type.ts
@@ -1,0 +1,5 @@
+export interface INavigateProps {
+  img: string
+  description: string
+  type: number
+}
